+++ packages/react/src/KeybindingHint/components/Sequence.tsx
@@ -1,8 +1,8 @@
 import {Fragment} from 'react'
 import type {KeybindingHintProps} from '../props'
 import VisuallyHidden from '../../_VisuallyHidden'
-import {accessibleChordString, Chord} from './Chord'
-import type {Platform} from '../platform'
+import {Chord} from './Chord'
+import {accessibleSequenceString} from './utils'
 
 const splitSequence = (sequence: string) => sequence.split(' ')
 
@@ -21,8 +21,4 @@ export const Sequence = ({keys, ...chordProps}: KeybindingHintProps) =>
     </Fragment>
   ))
 
-/** Plain string version of `Sequence` for use in `aria` string attributes. */
-export const accessibleSequenceString = (sequence: string, platform: Platform) =>
-  splitSequence(sequence)
-    .map(chord => accessibleChordString(chord, platform))
-    .join(' then ')
+export {accessibleSequenceString}
