+++ packages/react/src/Overlay/constants.ts
@@ -1,0 +1,19 @@
+export const heightMap = {
+  xsmall: '192px',
+  small: '256px',
+  medium: '320px',
+  large: '432px',
+  xlarge: '600px',
+  auto: 'auto',
+  initial: 'auto',
+  'fit-content': 'fit-content',
+}
+
+export const widthMap = {
+  small: '256px',
+  medium: '320px',
+  large: '480px',
+  xlarge: '640px',
+  xxlarge: '960px',
+  auto: 'auto',
+}
