+++ packages/react/src/Timeline/constants.ts
@@ -1,0 +1,11 @@
+export const TimelineBadgeVariants = [
+  'accent',
+  'success',
+  'attention',
+  'severe',
+  'danger',
+  'done',
+  'open',
+  'closed',
+  'sponsors',
+] as const
