+++ packages/react/src/Token/constants.ts
@@ -1,0 +1,10 @@
+export type TokenSizeKeys = 'small' | 'medium' | 'large' | 'xlarge'
+
+export const tokenSizes: Record<TokenSizeKeys, string> = {
+  small: '16px',
+  medium: '20px',
+  large: '24px',
+  xlarge: '32px',
+}
+
+export const defaultTokenSize: TokenSizeKeys = 'medium'
