+++ packages/react/src/UnderlineNav/utils.ts
@@ -1,0 +1,6 @@
+import React from 'react'
+
+export const getValidChildren = (children: React.ReactNode) => {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  return React.Children.toArray(children).filter(child => React.isValidElement(child)) as React.ReactElement<any>[]
+}
